project

+++ 02-JavaScript/11-projects/02-js-free-project/Eric/toDo.ts
@@ -12,16 +12,16 @@ class Item {
 
 class Items {
     allItems: Array<Item> = [];
-
     addItem(task: string): void {
+        try{
         let newItem: Item = new Item(task);
         this.allItems.push(newItem);
-        localStorage.setItem('item', JSON.stringify(this.allItems))
-
+        if (!this.allItems) throw new Error('The array where you want to push the task it doesn´t exist!');
+        localStorage.setItem('item', JSON.stringify(this.allItems));   
+    }catch(error){
+        console.error(error)
     }
-
-    
-
+    }
 }
 
 
@@ -29,15 +29,22 @@ const items = new Items();
 
 const doingSubmit = (ev: any): void => {
     ev.preventDefault();
-
+    try{
     const task: string = ev.target.elements.task.value;
-
     items.addItem(task);
+}catch(error){
+    console.error(error)
+}
 
 };
 
 function redirect() {
+    try{
     window.location.href = 'ItemsList.html'
+    if (!window.location.href) throw new Error('The page where you want to redirect it doesn´t exist!');
+    }catch(error){
+        console.error(error)
+    }
 }
 
 localStorage.clear()
